added a basic jump mechanic
added a basic jump. But it's buggy, need to work on the animation changes. Tends to hang in the air after jumping and consequtive jumps don't work so I've limited it to one jump per space bar press. Need to look into maybe casting a ray to the ground to switch back to normal animations to prevent hanging and fix consequtive jump problem.